fix: 修复 CI run 31875758026 剩余 4 条编译错误（3 文件）
- CardTextures: 纸纤维分叉角度 CGFloat+Double 混合表达式推断歧义致 cos 调用
  报 ambiguous；显式 Double(f.angle)/CGFloat(...) 统一类型链
- DatabaseRecoveryView: String(localized:) 的 LocalizationValue 插值不支持 Date，
  先 .formatted() 转 String（xcstrings key recovery.diagTime %@ 不变；
  .utf8 推断错误为连锁，随本行修复而解）
- EditorCanvasView: GraphicsContext.draw(_:in:slice:) 为 iOS 18 API（且 slice 为
  单位坐标，原代码传像素坐标本身有误）；iOS 17 改用 CGImage.cropping 取子图
  绘制，与导出路径（CGImage.cropping）一致

+++ MiLens/Views/Create/CardTextures.swift
@@ -131,9 +131,11 @@ struct PaperFibers: View {
             if rnd.next() < 0.18 {
                 var branch = Path()
                 branch.move(to: f.mid)
-                let a = f.angle + (rnd.next() < 0.5 ? 1 : -1) * (0.9 + rnd.next() * 0.7)
+                let a = Double(f.angle) + (rnd.next() < 0.5 ? 1 : -1) * (0.9 + rnd.next() * 0.7)
                 let len = 7 + rnd.next() * 9
-                branch.addLine(to: CGPoint(x: f.mid.x + cos(a) * len, y: f.mid.y + sin(a) * len))
+                branch.addLine(to: CGPoint(
+                    x: f.mid.x + CGFloat(cos(a) * len),
+                    y: f.mid.y + CGFloat(sin(a) * len)))
                 ctx.stroke(branch, with: .color(ink.opacity(0.26)),
                            style: StrokeStyle(lineWidth: 0.8, lineCap: .round))
             }
